Full refactor (#102) (#103)
This PR reflects a full rewrite of the code base. Key new features:

The program can now be shipped as a full python module. It can now be built and installed as a binary. This includes the use of a class structure for the device.
The main script has been broken up into more files. constants.py still retains the constants that are used. utilities.py now contains the helper functions such as device enumeration, launching the chimera app, etc. devices.py now contains the macro level device loops for managing input devices, including looping through events, discovery and capture of input devices, and creating the virtual device.
Individual handheld generations are now self contained in their own files. This will improve readability of the script and ensures adding new devices is less likely to break older devices that are less likely to be tested.
Devices that support momentary(instant) key-presses will now respond immediately, rather than waiting for an up event to trigger them. This should improve the support for the guide button on most devices that use a remapped guide button. AOKZOE/OXP and Ally devices do not support momentary buttons so they will not be affected.
Gyro support has been temporarily removed. The current state of gyro drivers in Linux is not very good, so being able to maintain support across the full range of devices is not currently possible. The old method was not well implemented in my opinion anyway, so I don't see this as a new negative at this time.